refactor to get endpoints in the same file named
