+++ app/weather/routes.py
